Skilt ut region-filteret i egen komponent
