+++ src/app/modules/side-menu/components/region-filter/region-filter.component.html
@@ -1,0 +1,47 @@
+<!-- Snøskredregioner -->
+<div *ngIf="regions$ | async as regions" slot="content">
+  <ion-accordion value="a-regions">
+    <ion-item slot="header" color="light">
+      <app-selected-items-counter-label [selectedItemsCount]="nRegionsSelected$ | async">
+        {{ "OBSERVATION_FILTER.REGION_FILTER_TITLE_SNOW" | translate }}
+      </app-selected-items-counter-label>
+    </ion-item>
+    <div slot="content">
+      <ion-list lines="none">
+        <ion-item *ngFor="let region of regions.a; trackBy: avalancheRegionTrackById">
+          <ion-checkbox
+            [value]="region"
+            [checked]="region.checked"
+            (ionChange)="regionCheckBoxChanged($event)"
+            labelPlacement="end"
+            justify="start"
+          >
+            {{ region.name }}
+          </ion-checkbox>
+        </ion-item>
+      </ion-list>
+    </div>
+  </ion-accordion>
+  <ion-accordion value="b-regions">
+    <ion-item slot="header" color="light">
+      <app-selected-items-counter-label [selectedItemsCount]="nRegionsSelected$ | async">
+        {{ "OBSERVATION_FILTER.REGION_FILTER_B_REGIONS_TITLE" | translate }}
+      </app-selected-items-counter-label>
+    </ion-item>
+    <div slot="content">
+      <ion-list lines="none">
+        <ion-item *ngFor="let region of regions.b; trackBy: avalancheRegionTrackById">
+          <ion-checkbox
+            [value]="region"
+            [checked]="region.checked"
+            (ionChange)="regionCheckBoxChanged($event)"
+            labelPlacement="end"
+            justify="start"
+          >
+            {{ region.name }}
+          </ion-checkbox>
+        </ion-item>
+      </ion-list>
+    </div>
+  </ion-accordion>
+</div>
